pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zapret-hub-mac"
version = "1.0.0b0"
description = "Native macOS rewrite of Zapret Hub with a macOS-only architecture."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "goshkow" }
]
dependencies = [
    "PySide6>=6.7",
    "cryptography>=46.0.0",
    "pyobjc-framework-Cocoa>=12.1",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.8",
]

[project.scripts]
zapret-hub-mac = "zapret_hub_mac.main:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
